+++ tests/appconfig_preferences_ex_test.py
@@ -1,18 +1,45 @@
 import pytest
-
-from nc_py_api import NextcloudException
-
 from gfixture import NC_APP
 
+from nc_py_api import NextcloudExceptionNotFound
 
 if NC_APP is None or "app_ecosystem_v2" not in NC_APP.capabilities:
     pytest.skip("app_ecosystem_v2 is not installed.", allow_module_level=True)
 
 
 @pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
+def test_cfg_ex_get_value_invalid(class_to_test):
+    with pytest.raises(ValueError):
+        class_to_test.get_value("")
+
+
+@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
+def test_cfg_ex_get_values_invalid(class_to_test):
+    assert class_to_test.get_values([]) == []
+    with pytest.raises(ValueError):
+        class_to_test.get_values([""])
+    with pytest.raises(ValueError):
+        class_to_test.get_values(["", "k"])
+
+
+@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
 def test_cfg_ex_set_empty_key(class_to_test):
-    with pytest.raises(NextcloudException):
+    with pytest.raises(ValueError):
         class_to_test.set("", "some value")
+
+
+@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
+def test_cfg_ex_delete_invalid(class_to_test):
+    class_to_test.delete([])
+    with pytest.raises(ValueError):
+        class_to_test.delete([""])
+    with pytest.raises(ValueError):
+        class_to_test.delete(["", "k"])
+
+
+@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
+def test_cfg_ex_get_default(class_to_test):
+    assert class_to_test.get_value("non_existing_key", default="alice") == "alice"
 
 
 @pytest.mark.parametrize("value", ("0", "1", "12 3", ""))
@@ -36,7 +63,20 @@ def test_cfg_ex_delete(class_to_test):
     assert class_to_test.get_value("test_key") is None
     class_to_test.delete("test_key")
     class_to_test.delete(["test_key"])
-    class_to_test.delete([])
+    with pytest.raises(NextcloudExceptionNotFound):
+        class_to_test.delete("test_key", not_fail=False)
+    with pytest.raises(NextcloudExceptionNotFound):
+        class_to_test.delete(["test_key"], not_fail=False)
+
+
+@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
+def test_cfg_ex_get(class_to_test):
+    class_to_test.delete(["test key", "test key2"])
+    assert len(class_to_test.get_values(["test key", "test key2"])) == 0
+    class_to_test.set("test key", "123")
+    assert len(class_to_test.get_values(["test key", "test key2"])) == 1
+    class_to_test.set("test key2", "123")
+    assert len(class_to_test.get_values(["test key", "test key2"])) == 2
 
 
 @pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
@@ -56,24 +96,10 @@ def test_cfg_ex_multiply_delete(class_to_test):
 @pytest.mark.parametrize("key", ("k", "k y", " "))
 @pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
 def test_cfg_ex_get_non_existing(key, class_to_test):
-    try:
-        class_to_test.delete(key)
-    except NextcloudException:
-        pass
+    class_to_test.delete(key)
     assert class_to_test.get_value(key) is None
     assert class_to_test.get_values([key]) == []
     assert len(class_to_test.get_values([key, "non_existing_key"])) == 0
-
-
-@pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))
-def test_cfg_ex_get(class_to_test):
-    class_to_test.delete(["test key", "test key2"])
-    assert len(class_to_test.get_values(["test key", "test key2"])) == 0
-    class_to_test.set("test key", "123")
-    assert len(class_to_test.get_values(["test key", "test key2"])) == 1
-    class_to_test.set("test key2", "123")
-    assert len(class_to_test.get_values(["test key", "test key2"])) == 2
-    assert len(class_to_test.get_values([])) == 0
 
 
 @pytest.mark.parametrize("class_to_test", (NC_APP.appconfig_ex_api, NC_APP.preferences_ex_api))