+++ tests/logs_test.py
@@ -1,10 +1,10 @@
-import pytest
-from unittest import mock
 from copy import deepcopy
+from unittest import mock
+
+import pytest
+from gfixture import NC_APP
 
 from nc_py_api import LogLvl, NextcloudException
-
-from gfixture import NC_APP
 
 if NC_APP is None or "app_ecosystem_v2" not in NC_APP.capabilities:
     pytest.skip("app_ecosystem_v2 is not installed.", allow_module_level=True)