+++ tests/weather_status_test.py
@@ -1,9 +1,8 @@
 import pytest
+from gfixture import NC_TO_TEST
 
 from nc_py_api import NextcloudException
 from nc_py_api.weather_status import WeatherLocationMode
-
-from gfixture import NC_TO_TEST
 
 
 @pytest.mark.parametrize("nc", NC_TO_TEST)