+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - `VERIFY_NC_CERTIFICATE` option.
 - `apps.ex_app_get_list` and `apps.ex_app_get_info` methods.
 - `files.download2stream` and `files.upload_stream` methods.
+- most of `FileAPI` can accept `FsNode` as a path.
 
 ### Changed